Fix text and tests
